PROJECT/requirements.txt
cassandra-driver==3.25.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
docopt==0.6.2
Flask==2.1.1
Flask-CassandraDB==0.0.1
geomet==0.2.1.post1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
joblib==1.1.0
MarkupSafe==2.1.1
nltk==3.7
numpy==1.22.3
pandas==1.4.2
pipreqs==0.4.11
python-dateutil==2.8.2
pytz==2022.1
regex==2022.3.15
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
tqdm==4.64.0
urllib3==1.26.9
uwsgidecorators==1.1.0
Werkzeug==2.1.1
wincertstore==0.2
yarg==0.1.9
zipp==3.8.0
